requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
_tflow_select=2.3.0=mkl
absl-py=0.10.0=py37hc8dfbb8_0
appdirs=1.4.3=py_1
astor=0.8.1=pyh9f0ad1d_0
audioread=2.1.8=py37hc8dfbb8_2
autopep8=1.5.4=py_0
blas=1.0=mkl
blinker=1.4=py_1
brotlipy=0.7.0=py37h4ab8f01_1000
ca-certificates=2020.7.22=0
cachetools=4.1.1=py_0
certifi=2020.6.20=py37_0
cffi=1.14.3=py37h26f1ce3_0
chardet=3.0.4=py37hc8dfbb8_1007
click=7.1.2=pyh9f0ad1d_0
cryptography=3.1=py37h26f1ce3_0
cycler=0.10.0=py_2
decorator=4.4.2=py_0
ffmpeg=4.3.1=ha925a31_0
ffmpeg-python=0.2.0=py_0
freetype=2.10.2=hd328e21_0
future=0.18.2=py37hc8dfbb8_1
gast=0.2.2=py_0
google-auth=1.21.2=py_0
google-auth-oauthlib=0.4.1=py_2
google-pasta=0.2.0=pyh8c360ce_0
grpcio=1.31.0=py37h554427f_0
h5py=2.10.0=nompi_py37hde23a51_104
hdf5=1.10.6=nompi_he0bbb20_101
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=vc14hc45fdbb_0
idna=2.10=pyh9f0ad1d_0
importlib-metadata=1.7.0=py37hc8dfbb8_0
intel-openmp=2019.4=245
joblib=0.16.0=py_0
jpeg=9d=he774522_0
keras-applications=1.0.8=py_1
keras-preprocessing=1.1.0=py_0
kiwisolver=1.2.0=py37heaa310e_0
libblas=3.8.0=14_mkl
libcblas=3.8.0=14_mkl
libflac=1.3.3=h6538335_0
liblapack=3.8.0=14_mkl
libmklml=2019.0.5=0
libogg=1.3.2=hfa6e2cd_1002
libpng=1.6.37=ha81a0f5_2
libprotobuf=3.13.0=h200bbdf_0
librosa=0.8.0=pyh9f0ad1d_0
libsndfile=1.0.29=ha925a31_0
libtiff=4.1.0=h885aae3_6
libvorbis=1.3.7=ha925a31_0
llvmlite=0.34.0=py37h8b575af_1
lz4-c=1.9.2=h62dcd97_2
markdown=3.2.2=py_0
matplotlib=3.3.1=0
matplotlib-base=3.3.1=py37hba9282a_0
mkl=2019.4=245
mkl-service=2.3.0=py37hfa6e2cd_0
norbert=0.2.1=py_0
numba=0.51.2=py37h414f9d2_0
numpy=1.19.1=py37h7cee19a_2
oauthlib=3.0.1=py_0
olefile=0.46=py_0
openssl=1.1.1g=he774522_1
opt_einsum=3.3.0=py_0
packaging=20.4=pyh9f0ad1d_0
pandas=1.1.2=py37h1834ac0_0
pillow=7.2.0=py37hc826c6e_1
pip=20.2.3=py_0
pooch=1.2.0=py_0
praat-parselmouth=0.3.3=pypi_0
prompt-toolkit=1.0.14=pypi_0
protobuf=3.13.0=py37h1834ac0_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycodestyle=2.6.0=py_0
pycparser=2.20=pyh9f0ad1d_2
pydub=0.24.1=pypi_0
pygments=2.7.1=pypi_0
pyinquirer=1.0.3=pypi_0
pyjwt=1.7.1=py_0
pyopenssl=19.1.0=py_1
pyparsing=2.4.7=pyh9f0ad1d_0
pyqt=5.9.2=py37ha878b3d_0
pyreadline=2.1=py37_1001
pysocks=1.7.1=py37hc8dfbb8_1
pysoundfile=0.10.2=py_1001
python=3.7.8=h60c2a47_1_cpython
python-dateutil=2.8.1=py_0
python_abi=3.7=1_cp37m
pytz=2020.1=pyh9f0ad1d_0
qt=5.9.7=vc14h73c81de_0
regex=2020.7.14=pypi_0
requests=2.24.0=pyh9f0ad1d_0
requests-oauthlib=1.3.0=pyh9f0ad1d_0
resampy=0.2.2=py_0
rsa=4.6=pyh9f0ad1d_0
scikit-learn=0.23.2=py37hdc70db3_0
scipy=1.5.2=py37h9439919_0
seaborn=0.10.1=py_0
setuptools=49.6.0=py37hc8dfbb8_1
simpleaudio=1.0.4=pypi_0
sip=4.19.24=py37ha925a31_0
six=1.15.0=pyh9f0ad1d_0
spleeter=1.5.3=py37hc8dfbb8_2
sqlite=3.33.0=he774522_0
tensorboard=2.3.0=py_0
tensorboard-plugin-wit=1.6.0=pyh9f0ad1d_0
tensorflow=1.15.0=mkl_py37h3789bd0_0
tensorflow-base=1.15.0=mkl_py37h190a33d_0
tensorflow-estimator=1.15.1=pyh2649769_0
termcolor=1.1.0=py_2
threadpoolctl=2.1.0=pyh5ca1d4c_0
tk=8.6.10=he774522_0
toml=0.10.1=py_0
tornado=6.0.4=py37hfa6e2cd_0
urllib3=1.25.10=py_0
vc=14.1=h869be7e_1
vs2015_runtime=14.16.27012=h30e32a0_2
wcwidth=0.2.5=pypi_0
werkzeug=0.16.1=py_0
wheel=0.35.1=pyh9f0ad1d_0
win_inet_pton=1.1.0=py37_0
wincertstore=0.2=py37_1003
wrapt=1.12.1=py37h8055547_1
xz=5.2.5=h62dcd97_1
zipp=3.1.0=py_0
zlib=1.2.11=h62dcd97_1009
zstd=1.4.5=h1f3a1b7_2
